refactor(tui): extract list-pane ranking and pin-sync cache from mod.rs
`mod.rs` is the highest-churn file in the codebase and mixes unrelated
concepts (Screen/KeyOutcome/mouse types) with self-contained clusters
that have existing names. Split two of them into their own modules:
`list_ranking` (cross-cutting dual-pane filter/rank/sort, not
List-screen-scoped) and `pin_sync` (the Pins presentation cache).
No behavior change.

Part of #313
